Renamed llm.chat to llm.server
